ver4

+++ client/src/main/java/ClientController.java
@@ -11,20 +11,20 @@ import javafx.stage.FileChooser;
 import model.Command;
 import model.commandMessage;
 import model.dataMessage;
+import org.apache.log4j.Logger;
+
+import java.awt.*;
 import java.io.*;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
-import java.util.List;
 import java.util.ResourceBundle;
 
 public class ClientController implements Initializable {
-    private static final String clientPath = "client/clientFiles";
+    static final Logger log = Logger.getLogger(String.valueOf(ClientController.class));
     private byte[] buffer = new byte[102400];
-    //    private int port = 8189;
-//    private String host = "localhost";
     private ClientNetwork network;
 
     @FXML
@@ -45,45 +45,39 @@ public class ClientController implements Initializable {
     public Button del;
     @FXML
     public Button quit;
-    @FXML
-    public ProgressBar progressBar;
-    @FXML
-    public ProgressIndicator progressIndicator;
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
-        List<String> clientFiles = null;
         setRegView();
 
-        //        Thread service = new Thread(() -> {
         network = ClientNetwork.getInstance(
                 message -> {
                     String msg = message.toString();
                     Platform.runLater(() ->
                     {
-                        if ((message instanceof commandMessage) &&
-                                ((((commandMessage) message).getCommand().startsWith(Command.REG_OK)) ||
-                                        (((commandMessage) message).getCommand().startsWith(Command.AUTH_OK)))
+                        if (message instanceof commandMessage) {
+                            String comMess = ((commandMessage) message).getCommand();
+                            if ((comMess.startsWith(Command.REG_OK)) || (comMess.startsWith(Command.AUTH_OK))) {
+                                Client.login = ((commandMessage) message).getLogin();
+                                setWorkView();
+                                Client.ps.setTitle(Client.title + " - пользователь: " + Client.login);
+                                network.write(new commandMessage(Command.GET_FILE_LIST, login.getText(), password.getText()));
+                            }
+                            if ((comMess.startsWith(Command.REG_NO)) || (comMess.startsWith(Command.AUTH_NO)))
+                                response.setText(((commandMessage) message).getComment());
 
-                        ) {
-                            setWorkView();
-                        } else {
-                            if ((message instanceof commandMessage) && (((commandMessage) message).getCommand().startsWith(Command.GET_FILE_LIST))) {
-                                listView.getItems().removeAll();
-                                listView.getItems().addAll(Arrays.asList(((commandMessage) message).getComment().split("\n")));
-                            } else {
-                                if (message instanceof dataMessage) {
-                                    saveFile((dataMessage) message);
-                                }
+                            if (comMess.startsWith(Command.GET_FILE_LIST)) {
+                                listView.getItems().clear();
+                                if (!((commandMessage) message).getComment().isEmpty())
+                                    listView.getItems().addAll(Arrays.asList(((commandMessage) message).getComment().split("\n")));
                             }
                         }
-                        response.setText(msg);
+
+                        if (message instanceof dataMessage)
+                            saveFile((dataMessage) message);
                     });
                 }
         );
-        //        });
-        //        service.setDaemon(true);
-        //        service.start();
     }
 
     @FXML
@@ -99,7 +93,6 @@ public class ClientController implements Initializable {
     }
 
     private boolean checkLogPass() {
-        //        todo можно добавить сложность пароля
         if (login.getText().isEmpty() || password.getText().isEmpty())
             return false;
         else return true;
@@ -109,13 +102,11 @@ public class ClientController implements Initializable {
     public void add(ActionEvent actionEvent) throws IOException {
         FileChooser dialog = new FileChooser();
         dialog.setTitle("Выбор файлов");
-        dialog.getExtensionFilters().addAll(
-                new FileChooser.ExtensionFilter("Все файлы", "*.*")
-        );
+        dialog.getExtensionFilters().addAll(new FileChooser.ExtensionFilter("Все файлы", "*.*"));
         File result = dialog.showOpenDialog(Client.ps);
         if (result != null) {
             listView.getItems().add(result.getName());
-            send(result);
+            upLoad(result);
         }
     }
 
@@ -126,10 +117,44 @@ public class ClientController implements Initializable {
         cm.setComment(fileName);
         network.write(cm);
         listView.getItems().remove(fileName);
+
+        Path fp = Paths.get(Client.fl, Client.login, fileName);
+        if (Files.exists(fp)) {
+            try {
+                Files.delete(fp);
+            } catch (IOException e) {
+                log.error(e);
+            }
+        }
     }
 
-    public void send(File result) throws IOException {
-        String fileName = result.getName(); //listView.getSelectionModel().getSelectedItem();
+    @FXML
+    public void quit(ActionEvent actionEvent) {
+        listView.getItems().clear();
+        Client.ps.setTitle(Client.title);
+        network.write(new commandMessage(Command.END, login.getText(), password.getText()));
+        setRegView();
+    }
+
+    @FXML
+    public void downLoad(MouseEvent mouseEvent) throws InterruptedException {
+        if (!listView.getItems().isEmpty()) {
+            commandMessage gf = new commandMessage(Command.GET_FILES, login.getText(), password.getText());
+            gf.setComment(listView.getSelectionModel().getSelectedItem());
+            network.write(gf);
+//        Thread.sleep(2000);
+            Path p = Paths.get(Client.fl, Client.login, gf.getComment());
+//        try {
+//            Desktop.getDesktop().open(new File(p.toString()));
+//            Desktop.getDesktop().open(new File(p.toString()));
+//        } catch (IOException e) {
+//            e.printStackTrace();
+//        }
+        }
+    }
+
+    public void upLoad(File result) throws IOException {
+        String fileName = result.getName();
         long len = Files.size(Paths.get(result.toString()));
 
         try (FileInputStream fis2 = new FileInputStream(result.getAbsolutePath())) {
@@ -147,6 +172,7 @@ public class ClientController implements Initializable {
 
                 dataMessage mess = new dataMessage(fileName, len);
                 mess.setContent(b2);
+                mess.setLogin(Client.login);
                 network.write(mess);
 
                 for (int i = 0; i < buffer.length; i++) {
@@ -154,19 +180,12 @@ public class ClientController implements Initializable {
                 }
             }
         } catch (IOException e) {
-            System.out.println("Ошибка чтения файла: " + fileName + " " + e);
+            log.info("Ошибка чтения файла: " + fileName + " " + e);
         }
     }
 
     public void clearResponse() {
         response.setText("");
-    }
-
-    @FXML
-    public void quit(ActionEvent actionEvent) {
-        setRegView();
-        //            commandMessage mess;
-////            mess = new commandMessage(Command.END, login.getText(), password.getText());
     }
 
     public void setRegView() {
@@ -181,8 +200,6 @@ public class ClientController implements Initializable {
         del.setVisible(false);
         quit.setVisible(false);
         auth.setVisible(true);
-        progressBar.setVisible(false);
-        progressIndicator.setVisible(false);
     }
 
     public void setWorkView() {
@@ -194,40 +211,35 @@ public class ClientController implements Initializable {
         listView.setVisible(true);
         add.setVisible(true);
         del.setVisible(true);
-//        progressBar.setVisible(true);
-//        progressIndicator.setVisible(true);
-        network.write(new commandMessage(Command.GET_FILE_LIST, login.getText(), password.getText()));
-    }
-
-    @FXML
-    public void downLoad(MouseEvent mouseEvent) throws InterruptedException {
-        commandMessage gf = new commandMessage(Command.GET_FILES, login.getText(), password.getText());
-        gf.setComment(listView.getSelectionModel().getSelectedItem());
-        network.write(gf);
-        Thread.sleep(2000);
-//        Path p = Paths.get(Client.fl + gf.getComment());
-//        try {
-//            Desktop.getDesktop().open(new File(Client.fl + gf.getComment()));
-//        } catch (IOException e) {
-//            e.printStackTrace();
-//        }
-        response.setText(gf.getComment());
     }
 
     public void saveFile(dataMessage mess) {
-        Path p = Paths.get(Client.fl + mess.getFileName());
-        try (FileOutputStream fos = new FileOutputStream(p.toString(), true)) {
-            fos.write(mess.getContent());
-        } catch (FileNotFoundException e) {
-            System.out.println(": " + e);
-//            log.info(": " + e);
-//            e.printStackTrace();
+        createDirIfNotExist(Client.login);
+
+        Path p = Paths.get(Client.fl, Client.login, mess.getFileName());
+        try {
+            if (Files.notExists(p) || Files.size(p) < mess.getFileSize()) {
+
+                try (FileOutputStream fos = new FileOutputStream(p.toString(), true)) {
+                    fos.write(mess.getContent());
+                } catch (FileNotFoundException e) {
+                    log.info(": " + e);
+                }
+            }
         } catch (IOException e) {
-//            e.printStackTrace();
-            System.out.println(": " + e);
-//            log.info(": " + e);
+            e.printStackTrace();
         }
     }
 
+    public static void createDirIfNotExist(String login) {
+        Path fdir = Paths.get(Client.fl, Client.login);
+        if (Files.notExists(fdir)) {
+            try {
+                Files.createDirectory(fdir);
+            } catch (IOException e) {
+                log.error(e);
+            }
+        }
+    }
 
 }
